moves sendMessengerAction(typing/seen msg) to util/messenger
