+++ apps/nextjs/src/util/messenger/sendMessengerAction.ts
@@ -1,0 +1,24 @@
+import axios from "axios";
+
+import { env } from "@brain2/ai";
+
+import type { SenderAction } from "./types/senderAction";
+import type { SenderActionRequest } from "./types/senderActionRequest";
+
+// Mark a message as seen or turn the typing animation on
+export async function sendMessengerAction(
+  senderPSID: string,
+  action: SenderAction,
+): Promise<void> {
+  const request: SenderActionRequest = {
+    recipient: { id: senderPSID },
+    sender_action: action,
+  };
+
+  await axios.post(env.MESSENGER_API_URL, request, {
+    headers: {
+      Authorization: `Bearer ${env.MESSENGER_ACCESS_TOKEN}`,
+      "Content-Type": "application/json",
+    },
+  });
+}
